refactor(wallet-integration): move to module with pop-cli crate
