--- crates/pop-wallet-integration/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "pop-wallet-integration"
-edition.workspace = true
-documentation.workspace = true
-license.workspace = true
-repository.workspace = true
-version.workspace = true
-
-[dependencies]
-axum.workspace = true
-tokio.workspace = true
-tower-http = { workspace = true, features = ["fs"] }
-serde_json.workspace = true
-serde = { workspace = true, features = ["derive"] }
